fix: replace plaintext formatter with jsonl
